+++ src/egttools/Data.cpp
@@ -16,10 +16,7 @@
   * along with EGTtools.  If not, see <http://www.gnu.org/licenses/>
 */
 
-#include <egttools/Types.h>
-
 #include <egttools/Data.hpp>
-#include <utility>
 
 egttools::DataStructures::DataTable::DataTable(size_t nb_rows, size_t nb_columns,
                                                std::vector<std::string> &headers, std::vector<std::string> &column_types)
@@ -27,5 +24,5 @@ egttools::DataStructures::DataTable::DataTable(size_t nb_rows, size_t nb_columns
       nb_columns(nb_columns),
       header(std::move(headers)),
       column_types(std::move(column_types)) {
-    data = Matrix2D::Zero(nb_rows, nb_columns);
+    data = Matrix2D::Zero(static_cast<long>(nb_rows), static_cast<long>(nb_columns));
 }
